update
vista listado, imagenes de partidos

+++ ciudadanos/listarciudadano.php
@@ -1,7 +1,7 @@
 <?php 
 
 include "../layout/layout.php";
- require_once '../database/servicio.php';
+require_once '../database/servicio.php';
 require_once "../database/Iserviciobase.php";
 require_once "ciudadanoservice.php";
 require_once "../database/Context.php";
@@ -12,12 +12,6 @@ require_once "ciudadano.php";
 $service = new ciudadanoservice("database");
 
 $listarciudadano = $service->Getlista();
-
-
-
-
-
-  
 
 ?>
        
@@ -32,55 +26,47 @@ $listarciudadano = $service->Getlista();
             </div>
             <?php printHeader(true); ?>
 
+            <div class="contenedor">
+                <center><h3>Listado de ciudadanos</h3></center>
+                <div class="row justify-content-center">
+                  <table class="table fill">
+                    <thead>
+                      <tr>
+                        <th>Documento de Identidad</th>
+                        <th>Nombre</th>
+                        <th>Apellido</th>
+                        <th>Email</th>
+                        <th>Estado</th>
+                        <th colspan="2">Acción</th>
+                      </tr>
+                    </thead>
+                    <tbody>
+                    <?php
 
-            <h3 class="font-weight-bold">Ciudadanos</h3>
-            <br>
-            <div class="row">
+                      if($listarciudadano != null)
+                      {
+                          foreach ($listarciudadano as $ciudadano) :        
+                    
+                    ?>
+                        <td><?=$ciudadano->Identidad?></td>
+                        <td><?=$ciudadano->Nombre?></td>
+                        <td><?=$ciudadano->Apellido?></td>
+                        <td><?=$ciudadano->Email?></td>
+                        <td><?php if ($ciudadano->Estado == 1) : echo "Activo"; else: echo "Inactivo"; endif; ?></td>
+                        <td>
+                          <a href="editarciudadano.php?id=<?php echo $ciudadano->Identidad; ?>" class="card-link btn btn-warning btn-block">Editar</a>
+                          <a href="eliminarciudadano.php?id=<?php echo $ciudadano->Identidad; ?>" class="card-link btn btn-danger btn-block ml-0" onclick="return confirm('¿Estas seguro de querer eliminar este ciudadano?')">Eliminar</a>
+                        </td>
+                      </tr>
 
-                <?php foreach ($listarciudadano as $ciudadano) : ?>
-             
-                <div class="card text-white bg-dark cover-container" style=" width: 15rem" ;>
-                    <div class="card-body">
-                        <h5 class="card-title">Cedula: <?php echo $ciudadano->Identidad?></h5>
-                        <h5 class="card-subtitle mb-2"><?php echo $ciudadano->Nombre?></h5>
-                        <h6 class="card-text"><?php echo $ciudadano->Apellido?></h6>
-                        <h6 class="card-text"><?php echo $ciudadano->Email?></h6>
-                        <h6 class="card-text" id="estado">Estado:<?php if ($ciudadano->Estado == 1): ?>
+                    <?php
+                      endforeach;
+                    }
 
-<td>Activo</td>
-<?php else: ?>
-
-    <td>Inactivo</td>
-
-<?php endif ?>
-                        </h6>  <a href="editarciudadano.php?id=<?php echo $ciudadano->Identidad; ?>" class="card-link btn btn-outline-primary">Editar</a>
-                        <a href="eliminarciudadano.php?id=<?php echo $ciudadano->Identidad; ?>"
-                            class="card-link btn btn-outline-danger" onclick="return confirmar()">Eliminar</a>
-
-</div>
-
-
-
-                    </div>
-                    &nbsp; &nbsp;
-                  
-                    <?php endforeach; ?>
-
+                    ?>
+                    </tbody>
+                  </table>
                 </div>
-                
-    
+              </div>
 
-            <?php printFooter(true); ?>
-
-            <script type="text/javascript">
-            function confirmar() {
-                var respuesta = confirm("Seguro de eliminar a este Ciudadano??");
-                if (respuesta == true) {
-                    return true;
-                } else {
-                    return false;
-                }
-            }
-    
-
-            </script>
+              <?php printFooter(true); ?>
